+++ boost/accumulators/statistics/tail_variate_means.hpp
@@ -33,41 +33,41 @@ namespace boost { namespace accumulators
 {
 
 namespace impl
-{       
+{
     /**
         @brief Estimation of the absolute and relative tail variate means (for both left and right tails)
 
-        For all \f$j\f$-th variates associated to the \f$\lceil n(1-\alpha)\rceil\f$ largest samples (or the 
-        \f$\lceil n(1-\alpha)\rceil\f$ smallest samples in case of the left tail), the absolute tail means 
-        \f$\widehat{ATM}_{n,\alpha}(X, j)\f$ are computed and returned as an iterator range. Alternatively, 
-        the relative tail means \f$\widehat{RTM}_{n,\alpha}(X, j)\f$ are returned, which are the absolute 
+        For all \f$j\f$-th variates associated to the \f$\lceil n(1-\alpha)\rceil\f$ largest samples (or the
+        \f$\lceil n(1-\alpha)\rceil\f$ smallest samples in case of the left tail), the absolute tail means
+        \f$\widehat{ATM}_{n,\alpha}(X, j)\f$ are computed and returned as an iterator range. Alternatively,
+        the relative tail means \f$\widehat{RTM}_{n,\alpha}(X, j)\f$ are returned, which are the absolute
         tail means normalized with the (non-coherent) sample tail mean \f$\widehat{NCTM}_{n,\alpha}(X)\f$.
-        
+
         \f[
-            \widehat{ATM}_{n,\alpha}^{\mathrm{right}}(X, j) = 
+            \widehat{ATM}_{n,\alpha}^{\mathrm{right}}(X, j) =
                 \frac{1}{\lceil n(1-\alpha) \rceil}
                 \sum_{i=\lceil \alpha n \rceil}^n \xi_{j,i}
         \f]
-        
+
         \f[
-            \widehat{ATM}_{n,\alpha}^{\mathrm{left}}(X, j) = 
+            \widehat{ATM}_{n,\alpha}^{\mathrm{left}}(X, j) =
                 \frac{1}{\lceil n\alpha \rceil}
                 \sum_{i=1}^{\lceil n\alpha \rceil} \xi_{j,i}
         \f]
-        
+
         \f[
-            \widehat{RTM}_{n,\alpha}^{\mathrm{right}}(X, j) = 
+            \widehat{RTM}_{n,\alpha}^{\mathrm{right}}(X, j) =
                 \frac{\sum_{i=\lceil n\alpha \rceil}^n \xi_{j,i}}
             {\lceil n(1-\alpha)\rceil\widehat{NCTM}_{n,\alpha}^{\mathrm{right}}(X)}
         \f]
-        
+
         \f[
-            \widehat{RTM}_{n,\alpha}^{\mathrm{left}}(X, j) = 
+            \widehat{RTM}_{n,\alpha}^{\mathrm{left}}(X, j) =
                 \frac{\sum_{i=1}^{\lceil n\alpha \rceil} \xi_{j,i}}
             {\lceil n\alpha\rceil\widehat{NCTM}_{n,\alpha}^{\mathrm{left}}(X)}
         \f]
     */
-    
+
     ///////////////////////////////////////////////////////////////////////////////
     // tail_variate_means_impl
     //  by default: absolute tail_variate_means
@@ -79,37 +79,37 @@ namespace impl
         typedef std::vector<float_type> array_type;
         // for boost::result_of
         typedef iterator_range<typename array_type::iterator> result_type;
-        
+
         tail_variate_means_impl(dont_care) {}
-        
+
         template<typename Args>
         result_type result(Args const &args) const
         {
             std::size_t cnt = count(args);
-        
+
             std::size_t n = static_cast<std::size_t>(
                 std::ceil(
                     cnt * ( ( is_same<LeftRight, left>::value ) ? args[quantile_probability] : 1. - args[quantile_probability] )
                 )
             );
-            
+
             std::size_t num_variates = tail_variate(args).begin()->size();
-            
+
             this->tail_means_.clear();
             this->tail_means_.resize(num_variates, Sample(0));
 
             // If n is in a valid range, return result, otherwise return NaN or throw exception
-            if (n < tail(args).size())
-            {     
+            if (n < static_cast<std::size_t>(tail(args).size()))
+            {
                 this->tail_means_ = std::accumulate(
                     tail_variate(args).begin()
                   , tail_variate(args).begin() + n
                   , this->tail_means_
                   , numeric::plus
                 );
-                
+
                 float_type factor = n * ( (is_same<Impl, relative>::value) ? non_coherent_tail_mean(args) : 1. );
-                
+
                 using boost::lambda::_1;
 
                 std::transform(
@@ -117,8 +117,8 @@ namespace impl
                   , this->tail_means_.end()
                   , this->tail_means_.begin()
                   , _1 / factor
-                );           
-            }  
+                );
+            }
             else
             {
                 if (std::numeric_limits<float_type>::has_quiet_NaN)
@@ -135,16 +135,16 @@ namespace impl
                     msg << "index n = " << n << " is not in valid range [0, " << tail(args).size() << ")";
                     boost::throw_exception(std::runtime_error(msg.str()));
                 }
-            }          
+            }
             return make_iterator_range(this->tail_means_);
         }
 
     private:
-    
+
         mutable array_type tail_means_;
-    
+
     };
-    
+
 } // namespace impl
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -159,7 +159,7 @@ namespace tag
     {
         typedef accumulators::impl::tail_variate_means_impl<mpl::_1, absolute, LeftRight, VariateTag> impl;
     };
-    template<typename LeftRight, typename VariateType, typename VariateTag>  
+    template<typename LeftRight, typename VariateType, typename VariateTag>
     struct relative_tail_variate_means
       : depends_on<count, non_coherent_tail_mean<LeftRight>, tail_variate<VariateType, VariateTag, LeftRight> >
     {
@@ -216,7 +216,7 @@ struct feature_of<tag::relative_tail_variate_means<LeftRight, VariateType, Varia
 {
 };
 
-// So that absolute_tail_means can be automatically substituted 
+// So that absolute_tail_means can be automatically substituted
 // with absolute_weighted_tail_means when the weight parameter is non-void.
 template<typename LeftRight, typename VariateType, typename VariateTag>
 struct as_weighted_feature<tag::absolute_tail_variate_means<LeftRight, VariateType, VariateTag> >
@@ -230,7 +230,7 @@ struct feature_of<tag::absolute_weighted_tail_variate_means<LeftRight, VariateTy
 {
 };
 
-// So that relative_tail_means can be automatically substituted 
+// So that relative_tail_means can be automatically substituted
 // with relative_weighted_tail_means when the weight parameter is non-void.
 template<typename LeftRight, typename VariateType, typename VariateTag>
 struct as_weighted_feature<tag::relative_tail_variate_means<LeftRight, VariateType, VariateTag> >
